move dummy function out
