+++ tests/ipython/dummy_function_fixtures.py
@@ -1,0 +1,8 @@
+def dummy_function(dummy_input, my_input):
+    """
+    Returns True if input is not
+    """
+    # this is an in-line comment in the body of the function
+    random_assignment = "Added for a longer function"
+    random_assignment += "make sure to modify variable"
+    return not dummy_input
